requirements.txt
enum34==1.1.6
piexif==1.1.3
Pillow==6.2.1